podcast_authority_study/requirements.txt
google-api-python-client>=2.100.0
google-auth>=2.23.0
youtube-transcript-api>=0.6.1
yt-dlp>=2024.1.0
openai-whisper>=20231117
pyyaml>=6.0
pandas>=2.1.0
tqdm>=4.66.0
anthropic>=0.40.0
tiktoken>=0.5.0
scikit-learn>=1.3.0
scipy>=1.11.0
